spa paging added

+++ BookApp.API/Dtos/BookCatalogItemDto.cs
@@ -2,9 +2,9 @@ using System;
 
 namespace BookApp.API.Dtos {
     public class BookCatalogItemDto {
-        public long Id { get; set; }
-        public long BookId { get; set; }
-        public long CatalogId { get; set; }
+        public int Id { get; set; }
+        public int BookId { get; set; }
+        public int CatalogId { get; set; }
         public string Name { get; set; }
         public bool IsPublic { get; set; }
         public DateTime Created { get; set; }
